[TAS-118] Refactor: 폴더명, 함수명, 클라스명 변경,
