--- lib/custom_map/components/const/data.dart
@@ -1,1 +1,0 @@
-const MY_PROFILE_IMAGE_PATH = 'assets/images/profile_pictures/user_profile.jpeg';
