+++ lib/custom_map/components/test_button/create_init_marker_button.dart
@@ -1,10 +1,10 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:frontend/custom_map/components/const/data.dart';
-import 'package:frontend/custom_map/components/const/type.dart';
-import 'package:frontend/custom_map/components/marker/custom_marker.dart';
-import 'package:frontend/custom_map/components/marker_icon.dart';
+import 'package:frontend/common/consts/data.dart';
+import 'package:frontend/custom_map/model/type.dart';
+import 'package:frontend/custom_map/components/marker/google_user_marker.dart';
+import 'package:frontend/custom_map/components/marker/user_marker_icon.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 class CreateInitMarkerButton extends StatelessWidget {
@@ -21,7 +21,7 @@ class CreateInitMarkerButton extends StatelessWidget {
         print("dfsdfdfdfdf");
         Random random = Random();
         int randomNumber = random.nextInt(100); // 0부터 99까지의 랜덤 숫자 생성
-        addMarker(await createMarker(
+        addMarker(await googleUserMarker(
           MarkerInfo(
             markerId: randomNumber.toString(),
             userName: '휘서',
