+++ lib/custom_map/model/type.dart
